# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FAMILIES IN FILM APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8082/api/",
            "description": "LOCAL API Server"
        },
        {
            "url": "https://fifapiqa.essentialdemo.com/api",
            "description": "LIVE API Server"
        }
    ],
    "paths": {
...
        "/partRequests": {
            "get": {
                "tags": [
                    "PartRequest"
                ],
                "summary": "Get a listing of the PartRequests.",
                "description": "Get all PartRequests",
                "operationId": "98b5d43bb08136f9766cf6b9c9d23573",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "PartRequest"
                ],
                "summary": "Store a newly created PartRequest in storage",
                "description": "Store PartRequest",
                "operationId": "4fc79ed9bac1dfc3034cd210bb55a9a0",
                "parameters": [
                    {
                        "name": "body",
                        "in": "path",
                        "description": "PartRequest that should be stored",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/partRequests/{id}": {
            "get": {
                "tags": [
                    "PartRequest"
                ],
                "summary": "Display the specified PartRequest",
                "description": "Get PartRequest",
                "operationId": "1829052d369aea7ae646941d232b7501",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PartRequest",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "PartRequest"
                ],
                "summary": "Update the specified PartRequest in storage",
                "description": "Update PartRequest",
                "operationId": "170254c0518a6295cc49967bbf1986fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PartRequest",
                        "required": true
                    },
                    {
                        "name": "body",
                        "in": "path",
                        "description": "PartRequest that should be updated",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PartRequest"
                ],
                "summary": "Remove the specified PartRequest from storage",
                "description": "Delete PartRequest",
                "operationId": "07705b4599a0e48c629cb6b4d9516140",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PartRequest",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}